Add working transaction filter by creating a filter subclass for both transaction and action lists/filters to inherit from. This includes repo, view model, and filter fragment
