+++ .idea/misc.xml
@@ -37,6 +37,7 @@
         <entry key="app/src/main/res/layout/fragment_actions.xml" value="0.25" />
         <entry key="app/src/main/res/layout/fragment_create.xml" value="0.1140625" />
         <entry key="app/src/main/res/layout/fragment_fill_variables.xml" value="0.2125" />
+        <entry key="app/src/main/res/layout/fragment_filter.xml" value="0.1586147926464301" />
         <entry key="app/src/main/res/layout/fragment_filter_actions.xml" value="0.2125" />
         <entry key="app/src/main/res/layout/fragment_record.xml" value="0.2125" />
         <entry key="app/src/main/res/layout/fragment_run_in_progress.xml" value="0.2125" />
@@ -56,6 +57,7 @@
         <entry key="app/src/main/res/layout/summary_graph.xml" value="0.30782385634886705" />
         <entry key="app/src/main/res/layout/transac_details_list_items.xml" value="0.24139492753623187" />
         <entry key="app/src/main/res/layout/transaction_details_fragment.xml" value="0.31657608695652173" />
+        <entry key="app/src/main/res/layout/transaction_filter_activity.xml" value="0.30782385634886705" />
         <entry key="app/src/main/res/layout/transaction_filter_fragment.xml" value="0.26666666666666666" />
         <entry key="app/src/main/res/layout/transaction_list_items.xml" value="0.24139492753623187" />
         <entry key="app/src/main/res/layout/transactions_status_summary.xml" value="0.33" />
