+++ app/src/main/java/com/hover/runner/transactions/TransactionsRepo.kt
@@ -1,10 +1,14 @@
 package com.hover.runner.transactions
 
 import androidx.lifecycle.LiveData
+import androidx.sqlite.db.SimpleSQLiteQuery
+import com.hover.runner.filters.FilterRepo
 import com.hover.sdk.database.HoverRoomDatabase
 import com.hover.sdk.transactions.Transaction
 
-class TransactionsRepo(private val sdkDB: HoverRoomDatabase) {
+class TransactionsRepo(private val sdkDB: HoverRoomDatabase): FilterRepo(sdkDB) {
+	override fun getTable(): String { return "SELECT * FROM hover_transactions" }
+
 	fun getAllTransactions(): LiveData<List<Transaction>> {
 		return sdkDB.transactionDao().allLive
 	}
@@ -27,5 +31,16 @@ class TransactionsRepo(private val sdkDB: HoverRoomDatabase) {
 
 	fun getLiveTransactionsByAction(actionId: String): LiveData<List<Transaction>> {
 		return sdkDB.transactionDao().getLiveTransactionsByActionId(actionId);
+	}
+
+	override fun search(sqlWhere: SimpleSQLiteQuery): List<Transaction> {
+		return sdkDB.transactionDao().search(sqlWhere)
+	}
+
+	override fun generateSearchString(search: String?): String {
+		if (search.isNullOrEmpty())
+			return ""
+		val s = "%$search%"
+		return "(uuid LIKE '$s' OR actionName LIKE '$s' OR actionId like '$s')"
 	}
 }
