+++ app/src/main/java/com/hover/runner/transactions/TransactionsViewModel.kt
@@ -1,10 +1,45 @@
 package com.hover.runner.transactions
 
+import android.app.Application
 import androidx.lifecycle.*
-import com.hover.sdk.actions.HoverAction
-import com.hover.sdk.parsers.HoverParser
+import androidx.sqlite.db.SimpleSQLiteQuery
+import com.hover.runner.actions.ActionRepo
+import com.hover.runner.filters.FilterViewModel
 import com.hover.sdk.transactions.Transaction
 
-class TransactionsViewModel(repo: TransactionsRepo) : ViewModel() {
-	val transactions: LiveData<List<Transaction>> = repo.getAllTransactions()
+class TransactionsViewModel(application: Application, actionRepo: ActionRepo, private val repo: TransactionsRepo) : FilterViewModel(application, actionRepo) {
+	val allTransactions: LiveData<List<Transaction>> = repo.getAllTransactions()
+	val filteredTransactions: MediatorLiveData<List<Transaction>> = MediatorLiveData()
+
+	init {
+		filteredTransactions.value = listOf()
+
+		filteredTransactions.apply {
+			addSource(allTransactions, this@TransactionsViewModel::runFilter)
+			addSource(filterQuery, this@TransactionsViewModel::runFilter)
+		}
+	}
+
+	private fun runFilter(transactions: List<Transaction>?) {
+		if (!transactions.isNullOrEmpty() && filterQuery.value == null) {
+			filteredTransactions.value = transactions
+			reset()
+		}
+	}
+
+	private fun runFilter(query: SimpleSQLiteQuery?) {
+		filteredTransactions.value = if (query != null) repo.search(query) else allTransactions.value
+	}
+
+	override fun onTransactionUpdate() {
+		TODO("Not yet implemented")
+	}
+
+	override fun generateSQLStatement(search: String?) {
+		search?.let { filterQuery.postValue(repo.generateSQLStatement(search, selectedTags.value)) }
+	}
+
+	override fun generateSQLStatement(tagList: List<String>?) {
+		tagList?.let {  filterQuery.postValue(repo.generateSQLStatement(searchString.value, tagList)) }
+	}
 }
