+++ app/src/main/res/layout/transaction_list_items.xml
@@ -7,42 +7,39 @@
     android:background="?selectableItemBackground"
     android:orientation="vertical">
 
-    <TextView
+    <TextView android:id="@+id/transaction_action_name"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:textStyle="bold"
-        android:id="@+id/transaction_action_name"
         android:textColor="@color/runnerWhite"
         android:drawablePadding="@dimen/margin8"
         android:paddingHorizontal="@dimen/margin16"
         android:textSize="@dimen/text_17"
         tools:text="Sample"/>
 
-    <TextView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:textStyle="bold"
-            android:textColor="@color/mainGrey"
-            android:id="@+id/transaction_date"
-            android:drawablePadding="@dimen/margin8"
-            android:paddingHorizontal="@dimen/margin16"
-            android:textSize="@dimen/text_13"
-            tools:text="March 3, 2020 at 5:43"/>
-    <TextView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:textColor="@color/mainGrey"
-            android:id="@+id/transaction_content"
-            android:layout_marginEnd="32dp"
-            android:paddingHorizontal="@dimen/margin16"
-            android:layout_marginBottom="16dp"
-            android:textSize="@dimen/text_15"
-            tools:text="This is a long message that could very well take multiple lines like three or four or five because it shows lots o details"/>
+    <TextView android:id="@+id/transaction_date"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:textStyle="bold"
+        android:textColor="@color/mainGrey"
+        android:drawablePadding="@dimen/margin8"
+        android:paddingHorizontal="@dimen/margin16"
+        android:textSize="@dimen/text_13"
+        tools:text="March 3, 2020 at 5:43"/>
 
-    <View
+    <TextView android:id="@+id/transaction_content"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:textColor="@color/mainGrey"
+        android:layout_marginEnd="32dp"
+        android:paddingHorizontal="@dimen/margin16"
+        android:layout_marginBottom="16dp"
+        android:textSize="@dimen/text_15"
+        tools:text="This is a long message that could very well take multiple lines like three or four or five because it shows lots o details"/>
+
+    <View android:id="@+id/divider"
         android:layout_width="match_parent"
         android:layout_height="1dp"
-        android:id="@+id/divider"
         android:background="@color/lineColor"/>
 
 </LinearLayout>
