+++ app/src/main/res/values/strings.xml
@@ -9,7 +9,7 @@
 
 	<string name="title_actions">Actions</string>
 	<string name="title_transactions">Transactions</string>
-	<string name="title_runs">Scheduled</string>
+	<string name="title_runs">Schedule</string>
 	<string name="title_run_details">Run Details</string>
 	<string name="title_settings">Settings</string>
 
@@ -109,7 +109,9 @@
 
 <!--	FILTERS-->
 	<string name="title_filter_actions">Filter Actions</string>
-	<string name="cta_filter_actions">Show %1$d Actions</string>
+	<string name="cta_filter_actions">Show %1$d Actions</string>\
+	<string name="title_filter_transactions">Filter Transactions</string>
+	<string name="cta_filter_transactions">Show %1$d Transactions</string>
 	<string name="reset"><u>Reset</u></string>
 	<string name="label_search">Search</string>
 	<string name="hint_search">Search for action ID or name</string>
@@ -128,7 +130,6 @@
 	<string name="failed_count_label">%1$d failed</string>
 	<string name="count_label">%1$d total</string>
 
-	<string name="title_filter_trans">Filter transactions</string>
 	<string name="search_hint_transaction">Search transaction id, or session text</string>
 	<string name="date_range">Date range</string>
 	<string name="countries">Countries</string>
